+++ CHANGELOG.md
@@ -60,8 +60,8 @@ messages inside `authz.MsgExec`.
 
 ### Improvements
 
-* #[131](https://github.com/babylonlabs-io/babylon/pull/131) Add new staking
-events
+* [#144](https://github.com/babylonlabs-io/babylon/pull/144) Add new finality provider events
+* [#131](https://github.com/babylonlabs-io/babylon/pull/131) Add new staking events
 * [#113](https://github.com/babylonlabs-io/babylon/pull/113) Add multibuild binary
 for upgrade handler `testnet` and `mainnet`.